# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OmegaBot platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /commands:
    get:
      operationId: listCommands
      tags: [commands]
      summary: List all commands
      responses:
        "200":
          description: Command list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandList"
    post:
      operationId: createCommand
      tags: [commands]
      summary: Issue a durable command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommandBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
  /commands/{id}:
    get:
      operationId: getCommand
      tags: [commands]
      summary: Get command by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
  /command-groups:
    get:
      operationId: listCommandGroups
      tags: [commands]
      summary: List command groups
      responses:
        "200":
          description: Command groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandGroupList"
    post:
      operationId: createCommandGroup
      tags: [commands]
      summary: Create command group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommandGroupBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandGroup"
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List approval requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, expired]
      responses:
        "200":
          description: Approvals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalList"
  /approvals/{id}/approve:
    post:
      operationId: approveRequest
      tags: [approvals]
      summary: Approve a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionBody"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{id}/reject:
    post:
      operationId: rejectRequest
      tags: [approvals]
      summary: Reject a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
...
  /github/change-plans:
    get:
      operationId: listChangePlans
      tags: [github]
      summary: List GitHub change plans
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Change plans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangePlanList"
    post:
      operationId: createChangePlan
      tags: [github]
      summary: Create a change plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChangePlanBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangePlan"
  /github/change-plans/{id}:
    get:
      operationId: getChangePlan
      tags: [github]
      summary: Get change plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Change plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangePlan"
...
components:
  schemas:
...
    Command:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [read, write, delete, external_call, webhook]
        status:
          type: string
          enum: [pending, executing, completed, failed, awaiting_approval, cancelled]
        isHighRisk:
          type: boolean
        requiresApproval:
          type: boolean
        idempotencyKey:
          type: string
        writeSafe:
          type: boolean
        groupId:
          type: string
        adapter:
          type: string
        createdAt:
          type: string
        executedAt:
          type: string
        payload:
          type: object
        result:
          type: object
      required: [id, name, type, status, createdAt]
    CommandList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Command"
        total:
          type: integer
      required: [items, total]
    CreateCommandBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        adapter:
          type: string
        isHighRisk:
          type: boolean
        requiresApproval:
          type: boolean
        idempotencyKey:
          type: string
        payload:
          type: object
      required: [name, type]
    CommandGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, partial]
        commandCount:
          type: integer
        completedCount:
          type: integer
        failedCount:
          type: integer
        createdAt:
          type: string
        commands:
          type: array
          items:
            $ref: "#/components/schemas/Command"
      required: [id, name, status, commandCount, createdAt]
    CommandGroupList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CommandGroup"
        total:
          type: integer
      required: [items, total]
    CreateCommandGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
    Approval:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        requestedBy:
          type: string
        requestedAt:
          type: string
        expiresAt:
          type: string
        decidedAt:
          type: string
        decidedBy:
          type: string
        reason:
          type: string
        commandId:
          type: string
        taskId:
          type: string
        adapter:
          type: string
        payload:
          type: object
      required: [id, title, status, riskLevel, requestedAt]
    ApprovalList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Approval"
        total:
          type: integer
        pendingCount:
          type: integer
      required: [items, total, pendingCount]
    ApprovalDecisionBody:
      type: object
      properties:
        reason:
          type: string
        decidedBy:
          type: string
...
    ChangePlan:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, pending_review, approved, rejected, applied, failed]
        repository:
          type: string
        branch:
          type: string
        prUrl:
          type: string
        prNumber:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        filesChanged:
          type: integer
        additions:
          type: integer
        deletions:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        reviewers:
          type: array
          items:
            type: string
        diff:
          type: string
        taskId:
          type: string
      required: [id, title, status, riskLevel, createdAt, updatedAt]
    ChangePlanList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChangePlan"
        total:
          type: integer
      required: [items, total]
    CreateChangePlanBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        repository:
          type: string
        branch:
          type: string
        taskId:
          type: string
        riskLevel:
          type: string
      required: [title, repository]
...